dotpatina::cli
Struct CliPatinaInterface

struct CliPatinaInterface {
    is_input_enabled: bool,
}

A struct for defining PatinaInterface behavior for the CLI.

Fields
is_input_enabled: bool

Whether or not input is enabled. When false, the CLI will not prompt for user confirmation and the CLI will continue until the action has completed or failed.

Implementations

impl CliPatinaInterface

fn new() -> CliPatinaInterface